# swagger.json
{"openapi":"3.0.0","paths":{"/users":{"get":{"operationId":"AppController_getUsers","summary":"Get Users List","parameters":[],"responses":{"200":{"description":"Get User List Success","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UsersResponseDto"}}}}}}},"post":{"operationId":"AppController_createUsers","summary":"Create User","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"Create Users Success"}}}},"/signin-line":{"post":{"operationId":"AppController_signInLine","summary":"SignIn Line","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInLineDto"}}}},"responses":{"200":{"description":"SignIn Line Success"}}}},"/logtime-notify":{"post":{"operationId":"AppController_lineNotify","summary":"Log Time Notify","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LineNotifyDto"}}}},"responses":{"200":{"description":"Log Time Notify Success"}}}}},"info":{"title":"JIRA Line Notify","description":"JIRA Line Notify API Description","version":"1.0.0","contact":{}},"tags":[{"name":"jita-line","description":""}],"servers":[],"components":{"schemas":{"UsersResponseDto":{"type":"object","properties":{"firstName":{"type":"string","example":"Anucha"},"lastName":{"type":"string","example":"Prapphrom"},"username":{"type":"string","example":"anucha"},"id":{"type":"string","example":"ZPg3I4tOivnNPcL9UUdd"}},"required":["firstName","lastName","username","id"]},"RegisterDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"}},"required":["username","password","firstName","lastName"]},"SignInLineDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"lineId":{"type":"string"}},"required":["username","password","lineId"]},"LineNotifyDto":{"type":"object","properties":{"userId":{"type":"string"},"logTimeHours":{"type":"number"}},"required":["userId","logTimeHours"]}}}}